Split typescript types files among specific components folder to follow map pattern
