+++ packages/visualizations/src/components/KpiCard/index.ts
@@ -1,6 +1,6 @@
 import KpiCardImpl from './KpiCard.svelte';
 import SvelteImpl from '../SvelteImpl';
-import type { KpiCardOptions } from '../types';
+import type { KpiCardOptions } from './types';
 
 export default class KpiCard extends SvelteImpl<number, KpiCardOptions> {
     protected getSvelteComponentClass(): typeof KpiCardImpl {
