+++ packages/visualizations/src/components/MarkdownText/types.ts
@@ -1,0 +1,3 @@
+export type MarkdownTextOptions = {
+    align?: 'left' | 'right' | 'center';
+};
